Closes-bug: #1420035 Fix unit test code compilation issues in trusty
'scons test flaky-test' found few issues when ran in trusty 14.04.
Some tests still time out, but with these changes, code should least
compile.

+++ src/io/test/tcp_server_test.cc
@@ -345,7 +345,7 @@ class ReaderTestSession : public TcpSession {
   private:
     bool ReceiveMsg(const u_int8_t *msg, size_t size) {
         TCP_UT_LOG_DEBUG("ReceiveMsg: " << size << " bytes");
-        if (size < reader_->GetHeaderLenSize())
+        if (size < (size_t) reader_->GetHeaderLenSize())
             return false;
         sizes.push_back(size);
         return true;
